// src/common/generated/prisma/enums.ts
export const OrganizationRole = {
  OWNER: 'OWNER',
  MEMBER: 'MEMBER'
} as const

export type OrganizationRole = (typeof OrganizationRole)[keyof typeof OrganizationRole]


export const TeamRole = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  DEVELOPER: 'DEVELOPER',
  VIEWER: 'VIEWER'
} as const

export type TeamRole = (typeof TeamRole)[keyof typeof TeamRole]
// ...
export const ActionEffect = {
  ALLOW: 'ALLOW',
  DENY: 'DENY'
} as const

export type ActionEffect = (typeof ActionEffect)[keyof typeof ActionEffect]


export const MatchType = {
  ALL: 'ALL',
  ANY: 'ANY'
} as const

export type MatchType = (typeof MatchType)[keyof typeof MatchType]


export const ApiKeyType = {
  CLIENT: 'CLIENT',
  SERVER: 'SERVER'
} as const

export type ApiKeyType = (typeof ApiKeyType)[keyof typeof ApiKeyType]


export const Operator = {
  EQUALS: 'EQUALS',
  IN: 'IN',
  INCLUDES: 'INCLUDES',
  GT: 'GT',
  LT: 'LT',
  GTE: 'GTE',
  LTE: 'LTE',
  CONTAINS: 'CONTAINS',
  STARTS_WITH: 'STARTS_WITH',
  ENDS_WITH: 'ENDS_WITH'
} as const

export type Operator = (typeof Operator)[keyof typeof Operator]
// ...
export const AuthProvider = {
  LOCAL: 'LOCAL',
  GITHUB: 'GITHUB'
} as const

export type AuthProvider = (typeof AuthProvider)[keyof typeof AuthProvider]